cleanup in finchers-core::output
