--- finchers-core/src/output/body.rs
@@ -1,141 +1,0 @@
-use bytes::Bytes;
-use futures::{self, Async, Stream};
-use hyper::body::Payload;
-use std::{fmt, io};
-
-use crate::future::Poll;
-
-#[allow(missing_docs)]
-#[derive(Debug)]
-pub struct Once<T>(Option<T>);
-
-impl<T: AsRef<[u8]> + Send + 'static> Payload for Once<T> {
-    type Data = io::Cursor<T>;
-    type Error = io::Error;
-
-    fn poll_data(&mut self) -> futures::Poll<Option<Self::Data>, Self::Error> {
-        Ok(Async::Ready(self.0.take().map(io::Cursor::new)))
-    }
-
-    fn is_end_stream(&self) -> bool {
-        self.0.is_none()
-    }
-
-    fn content_length(&self) -> Option<u64> {
-        self.0.as_ref().map(|body| body.as_ref().len() as u64)
-    }
-}
-
-#[allow(missing_docs)]
-pub fn once<T: AsRef<[u8]>>(data: T) -> Once<T> {
-    Once(Some(data))
-}
-
-/// An asynchronous stream representing the body of HTTP response.
-pub struct ResponseBody {
-    inner: Inner,
-}
-
-enum Inner {
-    Empty,
-    Once(Option<Bytes>),
-    Stream(Box<Stream<Item = Bytes, Error = io::Error> + Send>),
-}
-
-impl fmt::Debug for ResponseBody {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self.inner {
-            Inner::Empty => f.debug_tuple("Empty").finish(),
-            Inner::Once(ref bytes) => f.debug_tuple("Once").field(bytes).finish(),
-            Inner::Stream(..) => f.debug_tuple("Stream").finish(),
-        }
-    }
-}
-
-impl Default for ResponseBody {
-    fn default() -> ResponseBody {
-        ResponseBody::empty()
-    }
-}
-
-impl From<()> for ResponseBody {
-    fn from(_: ()) -> ResponseBody {
-        ResponseBody::empty()
-    }
-}
-
-macro_rules! impl_from_once {
-    ($($t:ty),*) => {$(
-        impl From<$t> for ResponseBody {
-            fn from(body: $t) -> ResponseBody {
-                ResponseBody::once(body)
-            }
-        }
-    )*};
-}
-
-impl_from_once!(&'static str, String, &'static [u8], Vec<u8>, Bytes);
-
-impl ResponseBody {
-    /// Create an instance of empty `ResponseBody`.
-    pub fn empty() -> ResponseBody {
-        ResponseBody {
-            inner: Inner::Empty,
-        }
-    }
-
-    /// Create an instance of `ResponseBody` from a chunk of bytes.
-    pub fn once<T>(body: T) -> ResponseBody
-    where
-        T: Into<Bytes>,
-    {
-        ResponseBody {
-            inner: Inner::Once(Some(body.into())),
-        }
-    }
-
-    /// Create an instance of `ResponseBody` from a `Stream` of bytes.
-    pub fn wrap_stream<T>(stream: T) -> ResponseBody
-    where
-        T: Stream<Item = Bytes, Error = io::Error> + Send + 'static,
-    {
-        ResponseBody {
-            inner: Inner::Stream(Box::new(stream)),
-        }
-    }
-
-    /// Return the length of bytes if available.
-    pub fn len(&self) -> Option<usize> {
-        match self.inner {
-            Inner::Empty => Some(0),
-            Inner::Once(ref chunk) => chunk.as_ref().map(|c| c.len()),
-            Inner::Stream(..) => None,
-        }
-    }
-
-    /// Poll an element of chunk from this stream.
-    pub fn poll_data(&mut self) -> Poll<io::Result<Option<Bytes>>> {
-        match self.inner {
-            Inner::Empty => Poll::Ready(Ok(None)),
-            Inner::Once(ref mut chunk) => Poll::Ready(Ok(chunk.take())),
-            Inner::Stream(ref mut stream) => match stream.poll() {
-                Ok(Async::Ready(chunk)) => Poll::Ready(Ok(chunk)),
-                Ok(Async::NotReady) => Poll::Pending,
-                Err(err) => Poll::Ready(Err(err)),
-            },
-        }
-    }
-}
-
-impl Payload for ResponseBody {
-    type Data = io::Cursor<Bytes>;
-    type Error = io::Error;
-
-    fn poll_data(&mut self) -> futures::Poll<Option<Self::Data>, Self::Error> {
-        match self.poll_data() {
-            Poll::Pending => Ok(Async::NotReady),
-            Poll::Ready(Ok(data)) => Ok(Async::Ready(data.map(io::Cursor::new))),
-            Poll::Ready(Err(err)) => Err(err),
-        }
-    }
-}
