+++ finchers-core/src/output/responders.rs
@@ -1,0 +1,7 @@
+#![allow(missing_docs)]
+
+mod debug;
+mod text;
+
+pub use self::debug::Debug;
+pub use self::text::Text;
